pyproject.toml
[tool.poetry]
name = ""
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "market"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.5"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.6"
ipython = "^8.14.0"
django-extensions = "^3.2.3"
dj-database-url = "^2.0.0"
django-jinja = "^2.10.2"
redis = "^5.0.0"
hiredis = "^2.2.3"
pillow = "^10.0.0"
django-debug-toolbar = "^4.2.0"
django-redis = "^5.3.0"
requests = "^2.31.0"
celery = "^5.3.4"
flower = "^2.0.1"
pydantic = "^2.4.2"
django-celery-beat = "^2.5.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.7.0"
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line_length = 119
